Updated windows commands to reflect recent changes
The current windows commands to setup a local registry no longer worked for me. The registry addon is setup to listen on port 80, not port 5000. It is unclear if this is a bug or intentional. If it is intentional, this new documentation should reflect this.

It changes the port forward from 5000:5000 to 5000:80, and adds another line to make it more clear that the "registry vm" is the registry service as obtained from the list of services.